Code review comments, registry.

+++ sdk/src/configuration/sdk_builder.cc
@@ -1219,7 +1219,7 @@ SdkBuilder::CreateBase2ExponentialBucketHistogramAggregation(
       std::make_unique<opentelemetry::sdk::metrics::Base2ExponentialHistogramAggregationConfig>();
 
   sdk->max_buckets_    = model->max_size;
-  sdk->max_scale_      = model->max_scale;
+  sdk->max_scale_      = static_cast<int32_t>(model->max_scale);
   sdk->record_min_max_ = model->record_min_max;
 
   return sdk;
